Move the offline alert tot he super class of the nav controller
